Controls append to .owl-wrapper-outer, not to the base item
I found out that using 'position: relative' on the element that activates the carousel cause a lot of strange floating bugs. This pull request will allow to include control arrows (prev/next) inside the item and will be relatively position to the ".item", not to the element that activates the carousel. And will allow you to set position: static to the base item, while the functionality and looks remain the same.

What possible problems may cause this, if not, merging this pull request will add more flexibility while not causing looks/functionality problems.